+++ packages/eslint-plugin/CHANGELOG.md
@@ -1,4 +1,4 @@
-# Next
+# 0.4.0
 - Added `semi-style`, `import/no-useless-path-segments`, `import/newline-after-import`, `import/extensions`, `react/jsx-max-props-per-line`, `react/no-typos`, `react/no-will-update-set-state`, `react/no-unused-prop-types`, `react/no-unused-state`, `react/no-render-return-value`, `react/no-this-in-sfc`, `react/no-access-state-in-setstate`, `react/void-dom-elements-no-children`
 - Update dependencies